Replace trading client with REST client
