+++ examples/cases/advanced/load_testing.py
@@ -5,11 +5,11 @@ from asyncio import run
 from typing import Set
 
 from examples.utils import BTC_USD_MARKET, create_trading_client
+from x10.clients.rest.rest_client import RestClient
 from x10.models.market import MarketModel
 from x10.models.order import OrderSide
 from x10.perpetual.order_object import create_order_object
 from x10.perpetual.stream_client.stream_client import PerpetualStreamClient
-from x10.perpetual.trading_client import PerpetualTradingClient
 
 LOGGER = logging.getLogger()
 MARKET_NAME = BTC_USD_MARKET
@@ -25,7 +25,7 @@ def generate_external_id():
     return str(random.randint(0, 10000000000000000000000000))
 
 
-async def create_orders_loop(*, trading_client: PerpetualTradingClient, market: MarketModel, level: int):
+async def create_orders_loop(*, trading_client: RestClient, market: MarketModel, level: int):
     market_mid_price = market.trading_config.round_price(
         (market.market_stats.bid_price + market.market_stats.ask_price) / 2
     )
@@ -81,7 +81,7 @@ async def order_confirmation_loop(*, stream_url: str, api_key: str):
                 pass
 
 
-async def cancel_open_orders(trading_client: PerpetualTradingClient):
+async def cancel_open_orders(trading_client: RestClient):
     positions = await trading_client.account.get_positions(market_names=[MARKET_NAME])
     balance = await trading_client.account.get_balance()
