+++ examples/cases/advanced/onboarding_with_eth_account.py
@@ -5,9 +5,9 @@ from eth_account import Account
 from eth_account.signers.local import LocalAccount
 
 from examples.utils import init_env
+from x10.clients.rest.rest_client import RestClient
 from x10.config import TESTNET_CONFIG
 from x10.core.stark_account import StarkPerpetualAccount
-from x10.perpetual.trading_client.trading_client import PerpetualTradingClient
 from x10.perpetual.user_client.user_client import UserClient
 from x10.utils.string import is_hex_string
 
@@ -36,7 +36,7 @@ async def run_example():
         private_key=main_account.l2_key_pair.private_hex,
         vault=main_account.account.l2_vault,
     )
-    trading_client = PerpetualTradingClient(CONFIG, starknet_account)
+    trading_client = RestClient(CONFIG, starknet_account)
 
     LOGGER.info("StarkNet public key: %s", starknet_account.public_key)
