+++ x10/clients/rest/__init__.py
@@ -1,0 +1,1 @@
+from x10.clients.rest.rest_client import RestClient  # noqa: F401
