+++ x10/clients/rest/modules/info_module.py
@@ -1,16 +1,35 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List, Optional
 
+from x10.clients.rest.modules.base_module import BaseModule
+from x10.models.asset import AssetModel
 from x10.models.candle import CandleInterval, CandleModel, CandleType
 from x10.models.funding_rate import FundingRateModel
 from x10.models.market import MarketModel, MarketStatsModel
 from x10.models.orderbook import OrderbookUpdateModel
-from x10.perpetual.trading_client.base_module import BaseModule
+from x10.models.settings import SettingsModel
 from x10.utils.date import to_epoch_millis
 from x10.utils.http import send_get_request
 
 
-class InfoMarketsModule(BaseModule):
+class InfoModule(BaseModule):
+    async def get_settings(self):
+        url = self._get_url("/info/settings")
+        return await send_get_request(await self.get_session(), url, SettingsModel)
+
+    async def get_assets(self):
+        url = self._get_url("/info/assets")
+        return await send_get_request(await self.get_session(), url, List[AssetModel])
+
+    async def get_assets_dict(self):
+        assets = await self.get_assets()
+        return {asset.name: asset for asset in assets.data}
+
+    async def get_asset_price(self, *, asset_name: str):
+        url = self._get_url("/info/assets/<asset_name>/price", asset_name=asset_name)
+        return await send_get_request(await self.get_session(), url, Decimal)
+
     async def get_markets(self, *, market_names: Optional[List[str]] = None):
         """
         https://api.docs.extended.exchange/#get-markets
