+++ x10/models/settings.py
@@ -1,0 +1,5 @@
+from x10.models.base import X10BaseModel
+
+
+class SettingsModel(X10BaseModel):
+    stark_ex_contract_address: str
