+++ x10/models/testnet.py
@@ -1,0 +1,5 @@
+from x10.models.base import X10BaseModel
+
+
+class ClaimResponseModel(X10BaseModel):
+    id: int
