--- x10/perpetual/trading_client/__init__.py
@@ -1,3 +1,0 @@
-from x10.perpetual.trading_client.trading_client import (  # noqa: F401
-    PerpetualTradingClient,
-)
