+++ x10/perpetual/user_client/onboarding.py
@@ -6,19 +6,11 @@ from eth_account.signers.local import LocalAccount
 from fast_stark_crypto import generate_keypair_from_eth_signature, pedersen_hash
 from fast_stark_crypto import sign as stark_sign
 
-from x10.models.account import AccountModel
-from x10.models.base import X10BaseModel
-
 register_action = "REGISTER"
 sub_account_action = "CREATE_SUB_ACCOUNT"
 
 
-class OnboardedClientModel(X10BaseModel):
-    l1_address: str
-    default_account: AccountModel
-
-
-@dataclass
+@dataclass(frozen=True)
 class StarkKeyPair:
     private: int
     public: int
@@ -32,7 +24,7 @@ class StarkKeyPair:
         return hex(self.private)
 
 
-@dataclass
+@dataclass(frozen=True)
 class AccountRegistration:
     account_index: int
     wallet: str
@@ -88,7 +80,7 @@ class AccountRegistration:
         }
 
 
-@dataclass
+@dataclass(frozen=True)
 class SubAccountOnboardingPayload:
     l2_key: int
     l2_r: int
@@ -108,7 +100,7 @@ class SubAccountOnboardingPayload:
         }
 
 
-@dataclass
+@dataclass(frozen=True)
 class OnboardingPayLoad:
     l1_signature: str
     l2_key: int
